moved constants into utils folder
